refactor(desktop): split oversized stability owners
Verified: product line-count ratchet targets are below baseline and 64 focused telemetry/realtime tests pass.
