+++ .github/scripts/product_file_line_count_ratchet_baseline/desktop-swift-root.json
@@ -4,7 +4,7 @@
     "desktop/macos/Desktop/Sources/CloudConnectorFormAutomation.swift": 1678,
     "desktop/macos/Desktop/Sources/DesktopAutomationBridge.swift": 4256,
     "desktop/macos/Desktop/Sources/MemoryExportService.swift": 1593,
-    "desktop/macos/Desktop/Sources/OmiApp.swift": 1693
+    "desktop/macos/Desktop/Sources/OmiApp.swift": 1635
   },
   "raise_justifications": {
     "desktop/macos/Desktop/Sources/AuthService.swift": "Apple first-auth name capture signals observers (authNameDidUpdate) and persists the name to Firebase so it survives reinstalls.",
