+++ desktop/macos/Desktop/Sources/FloatingControlBar/RealtimeHubController+SessionLifecycle.swift
@@ -1028,66 +1028,6 @@ extension RealtimeHubController {
     schedulePhysicalSessionTeardown(detachedSession)
   }
 
-  /// Replaces the current physical transport without overlap. Detachment is
-  /// synchronous (so stale callbacks are fenced immediately), but a new socket
-  /// cannot become visible until the old transport queue has completed close.
-  func replaceSessionAfterDrain(
-    preservingReconnectAudio: Bool = false,
-    preservingBargeInReplacement: Bool = false,
-    reconnectDelayNanoseconds: UInt64 = 0,
-    rewarmAfterDrain: Bool = true
-  ) {
-    guard !sessionReplacementGate.isPending else {
-      log("RealtimeHub: coalescing physical replacement while transport drain is pending")
-      return
-    }
-    let detachedSession = detachPhysicalSessionForTeardown(
-      preservingReconnectAudio: preservingReconnectAudio,
-      preservingBargeInReplacement: preservingBargeInReplacement)
-    let ownerGeneration = ownerBoundaryGeneration
-    let detachedSessionID = detachedSession.map(ObjectIdentifier.init)
-    if let detachedSession, let detachedSessionID {
-      detachedSessionsAwaitingDrain[detachedSessionID] = detachedSession
-    }
-    sessionReplacementGate.replace(
-      reconnectDelayNanoseconds: reconnectDelayNanoseconds,
-      stop: { [weak self, weak detachedSession] in
-        guard let self else { return }
-        if let detachedSession {
-          await detachedSession.stopAndWait()
-        }
-        if let detachedSessionID {
-          self.detachedSessionsAwaitingDrain.removeValue(forKey: detachedSessionID)
-        }
-      },
-      start: { [weak self] in
-        guard let self, self.ownerBoundaryGeneration == ownerGeneration else { return }
-        self.reconnectPending = false
-        let abandonedBargeInReplacement =
-          preservingBargeInReplacement && self.pendingBargeInOwnerScope == nil
-        if rewarmAfterDrain || abandonedBargeInReplacement, self.session == nil {
-          #if DEBUG
-            if let testingWarmAfterDrain = self.testingWarmAfterDrain {
-              testingWarmAfterDrain()
-              return
-            }
-          #endif
-          self.ensureWarm()
-        }
-      })
-  }
-
-  func schedulePhysicalSessionTeardown(_ detachedSession: RealtimeHubSession) {
-    let sessionID = ObjectIdentifier(detachedSession)
-    guard detachedSessionsAwaitingDrain[sessionID] == nil else { return }
-    detachedSessionsAwaitingDrain[sessionID] = detachedSession
-    Task { @MainActor [weak self, weak detachedSession] in
-      guard let detachedSession else { return }
-      await detachedSession.stopAndWait()
-      self?.detachedSessionsAwaitingDrain.removeValue(forKey: sessionID)
-    }
-  }
-
   func clearBargeInReplacementState() {
     bargeInReplacementGeneration &+= 1
     replacementAudioBuffer = nil
